LayoutTests/fast/frames/seamless/seamless-quirks-expected.txt

Test that a quirks-mode document inside a seamless iframe sizes correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframe.seamless is true
PASS iframe.contentDocument.compatMode is "BackCompat"
PASS window.getComputedStyle(iframe).width is "200px"
PASS window.getComputedStyle(iframe).height is "100px"
